[refactor] order 4계층으로 변경
