+++ order/src/main/java/com/sparta/rooibos/order/application/service/feign/DeliveryService.java
@@ -3,7 +3,7 @@ package com.sparta.rooibos.order.application.service.feign;
 import com.sparta.rooibos.order.application.dto.request.CreateDeliveryRequest;
 import com.sparta.rooibos.order.application.dto.request.UpdateDeliveryRequest;
 import com.sparta.rooibos.order.application.dto.response.CreateDeliveryResponse;
-import com.sparta.rooibos.order.infrastructure.delivery.UpdateDeliveryResponse;
+import com.sparta.rooibos.order.application.feigin.dto.UpdateDeliveryResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
